+++ CS350_BaggingApplication/Models/Item.cs
@@ -10,11 +10,11 @@ namespace CS350_BaggingApplication.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Length { get; set; }
-        public int Height { get; set; }
-        public int Width { get; set; }
-        public int Weight { get; set; }
+        public double Length { get; set; }
+        public double Height { get; set; }
+        public double Width { get; set; }
+        public double Weight { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
     }
 }
