+++ CS350_BaggingApplication/ViewModels/CalculateOrderResultViewModel.cs
@@ -10,7 +10,7 @@ namespace CS350_BaggingApplication.ViewModels
     {
         public Dictionary<Item, int> Items { get; set; }
         public Packaging PackagingUsed { get; set; }
-        public int TotalWeight { get; set; }
+        public double TotalWeight { get; set; }
         public int TotalItems { get; set; }
         public int BagsNeeded { get; set; }
     }
